+++ pkg/schemautils/schemautils.go
@@ -34,12 +34,12 @@ func (schema *ExtendedSchema) GetRefSchema() *ExtendedSchema {
 	return refSchema
 }
 
-func (schema *ExtendedSchema) GetProperties() []*ExtendedSchema {
-	properties := make([]*ExtendedSchema, 0, len(schema.Properties))
-	for _, property := range schema.Properties {
+func (schema *ExtendedSchema) GetProperties() map[string]*ExtendedSchema {
+	properties := map[string]*ExtendedSchema{}
+	for key, property := range schema.Properties {
 		newSchema := NewExtendedSchema(property)
 		newSchema.InheritParentFrom(schema)
-		properties = append(properties, newSchema)
+		properties[key] = newSchema
 	}
 	return properties
 }
@@ -164,23 +164,23 @@ func locationIsProperty(resolvedLocation string) bool {
 }
 
 func (schema *ExtendedSchema) IsObject() bool {
-	return schema.isType("object")
+	return schema.IsType("object")
 }
 
 func (schema *ExtendedSchema) IsArray() bool {
-	return schema.isType("array")
+	return schema.IsType("array")
 }
 
 func (schema *ExtendedSchema) IsString() bool {
-	return schema.isType("string")
+	return schema.IsType("string")
 }
 
 func (schema *ExtendedSchema) IsNumber() bool {
-	return schema.isType("number")
+	return schema.IsType("number")
 }
 
 func (schema *ExtendedSchema) IsInteger() bool {
-	return schema.isType("integer")
+	return schema.IsType("integer")
 }
 
 func (schema *ExtendedSchema) IsNumeric() bool {
@@ -188,10 +188,10 @@ func (schema *ExtendedSchema) IsNumeric() bool {
 }
 
 func (schema *ExtendedSchema) IsBoolean() bool {
-	return schema.isType("boolean")
+	return schema.IsType("boolean")
 }
 
-func (schema *ExtendedSchema) isType(typeName string) bool {
+func (schema *ExtendedSchema) IsType(typeName string) bool {
 	isType := false
 	for _, t := range schema.Types {
 		if t == typeName {